requirements.txt
selenium
requests
pyautogui
cryptography
pillow
flet>=0.21.0
winshell>=0.6
pywin32>=306